convert getVersion to long

+++ src/main/java/org/vvcephei/occ_map/Versioned.java
@@ -1,5 +1,5 @@
 package org.vvcephei.occ_map;
 
 public interface Versioned {
-    int getVersion();
+    long getVersion();
 }
